+++ libnmstate/nm/device.py
@@ -22,15 +22,12 @@ import logging
 from libnmstate.error import NmstateLibnmError
 
 from . import active_connection as ac
-from . import connection
+from . import profile
 
 
-def activate(context, dev=None, connection_id=None):
-    """Activate the given device or remote connection profile."""
-    conn = connection.ConnectionProfile(context)
-    conn.nmdevice = dev
-    conn.con_id = connection_id
-    conn.activate()
+def activate(context, dev=None, profile=None):
+    """Activate the given remote connection profile."""
+    profile.activate()
 
 
 def deactivate(context, dev):
@@ -49,7 +46,7 @@ def deactivate(context, dev):
 def delete(context, dev):
     connections = dev.get_available_connections()
     for con in connections:
-        con_profile = connection.ConnectionProfile(context, con)
+        con_profile = profile.Profile(context, con)
         con_profile.delete()
 
 
@@ -60,12 +57,12 @@ def modify(context, nm_dev, connection_profile):
     connection profile activation.
     """
     nm_ac = nm_dev.get_active_connection()
-    if connection.is_activated(nm_ac, nm_dev):
+    if profile.is_activated(nm_ac, nm_dev):
         version_id = 0
         flags = 0
         action = f"Reapply device config: {nm_dev.get_iface()}"
         context.register_async(action)
-        user_data = context, nm_dev, action
+        user_data = context, nm_dev, action, connection_profile
         nm_dev.reapply_async(
             connection_profile,
             version_id,
@@ -75,11 +72,11 @@ def modify(context, nm_dev, connection_profile):
             user_data,
         )
     else:
-        _activate_async(context, nm_dev)
+        _activate_async(context, nm_dev, connection_profile)
 
 
 def _modify_callback(src_object, result, user_data):
-    context, nmdev, action = user_data
+    context, nmdev, action, con_profile = user_data
     if context.is_cancelled():
         return
     devname = src_object.get_iface()
@@ -93,7 +90,7 @@ def _modify_callback(src_object, result, user_data):
             e,
         )
         context.finish_async(action, suppress_log=True)
-        _activate_async(context, src_object)
+        _activate_async(context, src_object, con_profile)
         return
 
     if success:
@@ -105,12 +102,13 @@ def _modify_callback(src_object, result, user_data):
             devname,
         )
         context.finish_async(action, suppress_log=True)
-        _activate_async(context, src_object)
+        _activate_async(context, src_object, con_profile)
 
 
-def _activate_async(context, dev):
-    conn = connection.ConnectionProfile(context)
+def _activate_async(context, dev, connection_profile):
+    conn = profile.Profile(context)
     conn.con_id = dev.get_iface()
+    conn.uuid = connection_profile.get_uuid()
     conn.nmdevice = dev
     if dev:
         # Workaround of https://bugzilla.redhat.com/show_bug.cgi?id=1772470
