+++ tests/integration/nm/bond_test.py
@@ -115,14 +115,14 @@ def _create_bond(ctx, name, options):
     ipv4_setting = nm.ipv4.create_setting({}, None)
     ipv6_setting = nm.ipv6.create_setting({}, None)
 
-    con_profile = nm.connection.ConnectionProfile(ctx)
+    con_profile = nm.profile.Profile(ctx)
     con_profile.create(
         (con_setting.setting, bond_setting, ipv4_setting, ipv6_setting)
     )
     with main_context(ctx):
         con_profile.add()
         ctx.wait_all_finish()
-        nm.device.activate(ctx, connection_id=name)
+        nm.device.activate(ctx, profile=con_profile)
 
 
 def _delete_bond(ctx, devname):
@@ -135,17 +135,17 @@ def _delete_bond(ctx, devname):
 
 def _attach_slave_to_bond(ctx, bond, slave):
     slave_nmdev = ctx.get_nm_dev(slave)
-    curr_slave_con_profile = nm.connection.ConnectionProfile(ctx)
+    curr_slave_con_profile = nm.profile.Profile(ctx)
     curr_slave_con_profile.import_by_device(slave_nmdev)
 
-    slave_con_profile = nm.connection.ConnectionProfile(ctx)
+    slave_con_profile = nm.profile.Profile(ctx)
     slave_settings = [_create_connection_setting(bond, curr_slave_con_profile)]
     slave_con_profile.create(slave_settings)
 
     with main_context(ctx):
         curr_slave_con_profile.update(slave_con_profile)
         ctx.wait_all_finish()
-        nm.device.activate(ctx, connection_id=slave)
+        nm.device.activate(ctx, profile=slave_con_profile)
 
 
 def _create_connection_setting(bond, port_con_profile):
