+++ tests/lib/nm/dns_test.py
@@ -1,5 +1,4 @@
-#
-# Copyright (c) 2019 Red Hat, Inc.
+# Copyright (c) 2019-2020 Red Hat, Inc.
 #
 # This file is part of nmstate
 #
@@ -21,10 +20,10 @@ from unittest import mock
 
 import pytest
 
-import libnmstate.nm.connection as nm_connection
 import libnmstate.nm.dns as nm_dns
 import libnmstate.nm.ipv4 as nm_ipv4
 import libnmstate.nm.ipv6 as nm_ipv6
+import libnmstate.nm.profile as nm_profile
 from libnmstate.ifaces import BaseIface
 from libnmstate.dns import DnsState
 from libnmstate.schema import DNS
@@ -131,7 +130,7 @@ def test_clear_dns(client_mock, nm_ip, get_test_dns_func):
         {InterfaceIP.ENABLED: True, BaseIface.DNS_METADATA: dns_conf},
         base_con_profile=None,
     )
-    con_profile = nm_connection.ConnectionProfile(client_mock)
+    con_profile = nm_profile.Profile(client_mock)
     con_profile.create([setting_ip])
     new_setting_ip = nm_ip.create_setting(
         {InterfaceIP.ENABLED: True, BaseIface.DNS_METADATA: {}},
