+++ tests/lib/nm/ovs_test.py
@@ -37,8 +37,8 @@ def context_mock():
 
 
 @pytest.fixture
-def nm_connection_mock():
-    with mock.patch.object(nm.ovs, "connection") as m:
+def nm_profile_mock():
+    with mock.patch.object(nm.ovs, "profile") as m:
         yield m
 
 
@@ -57,9 +57,9 @@ def test_is_ovs_interface_type_id(NM_mock):
     assert nm.ovs.is_ovs_interface_type_id(type_id)
 
 
-def test_get_ovs_info_without_ports(context_mock, nm_connection_mock, NM_mock):
+def test_get_ovs_info_without_ports(context_mock, nm_profile_mock, NM_mock):
     bridge_device = mock.MagicMock()
-    _mock_port_profile(nm_connection_mock)
+    _mock_port_profile(nm_profile_mock)
 
     device_info = [(bridge_device, None)]
     info = nm.ovs.get_ovs_info(context_mock, bridge_device, device_info)
@@ -77,12 +77,12 @@ def test_get_ovs_info_without_ports(context_mock, nm_connection_mock, NM_mock):
 
 
 def test_get_ovs_info_with_ports_without_interfaces(
-    nm_connection_mock, NM_mock, context_mock
+    nm_profile_mock, NM_mock, context_mock
 ):
     bridge_device = mock.MagicMock()
     port_device = mock.MagicMock()
-    _mock_port_profile(nm_connection_mock)
-    active_con = nm_connection_mock.get_device_active_connection.return_value
+    _mock_port_profile(nm_profile_mock)
+    active_con = nm_profile_mock.get_device_active_connection.return_value
     active_con.props.master = bridge_device
 
     device_info = [(bridge_device, None), (port_device, None)]
@@ -101,15 +101,15 @@ def test_get_ovs_info_with_ports_without_interfaces(
 
 
 def test_get_ovs_info_with_ports_with_interfaces(
-    nm_connection_mock, NM_mock, context_mock
+    nm_profile_mock, NM_mock, context_mock
 ):
     bridge_device = mock.MagicMock()
     port_device = mock.MagicMock()
     bridge_active_con = mock.MagicMock()
     port_active_con = mock.MagicMock()
     context_mock.get_nm_dev.return_value = port_device
-    _mock_port_profile(nm_connection_mock)
-    nm_connection_mock.get_device_active_connection = (
+    _mock_port_profile(nm_profile_mock)
+    nm_profile_mock.get_device_active_connection = (
         lambda dev: bridge_active_con
         if dev == bridge_device
         else port_active_con
@@ -175,8 +175,8 @@ def test_create_port_setting(NM_mock):
     assert port_setting.props.vlan_mode == vlan_mode
 
 
-def _mock_port_profile(nm_connection_mock):
-    con_prof_mock = nm_connection_mock.ConnectionProfile.return_value
+def _mock_port_profile(nm_profile_mock):
+    con_prof_mock = nm_profile_mock.Profile.return_value
     connection_profile = con_prof_mock.profile
     bridge_setting = connection_profile.get_setting.return_value
     bridge_setting.props.stp_enable = False
